♻️ Refactor: 컴포넌트에서 유틸 함수 분리 후 타입 정리
refactor: 컴포넌트에서 유틸 함수 분리 후 타입 정리
